requirements.txt
aiogram==3.13.1
tortoise-orm[aiosqlite]==0.21.6
python-dotenv==1.0.1
pandas==2.2.3
matplotlib==3.9.2
apscheduler==3.10.4